code-examples/LLM-Applications/FastAPI-RAG-LangGraph/pyproject.toml
[project]
name = "ai-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "chromadb>=0.6.3",
    "fastapi>=0.115.12",
    "greenlet>=3.2.3",
    "ipykernel>=6.29.5",
    "langgraph>=0.4.8",
    "openai>=1.85.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "pygraphviz>=1.14",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
]
